server/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
mysqlclient==2.1.0
Pillow==9.1.0
django-debug-toolbar==4.2.0
pylint-django==2.5.5
coverage==7.4.0
django-environ==0.11.2
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
gunicorn==21.2.0
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.1
